Move http client into web
